agent/requirements.txt
# Insurance Voice AI Agent - Python Dependencies

# Pipecat - Voice AI framework
pipecat-ai[daily,silero,deepgram,elevenlabs,anthropic]==0.0.44

# Twilio for telephony
twilio>=8.0.0

# Anthropic Claude API
anthropic>=0.21.0

# Deepgram for STT
deepgram-sdk>=3.0.0

# ElevenLabs for TTS
elevenlabs>=1.0.0

# Async support
aiohttp>=3.9.0
websockets>=12.0

# Environment and config
python-dotenv>=1.0.0

# Logging and debugging
loguru>=0.7.0

# Data handling
pydantic>=2.0.0
